improve logging

+++ resources/views/changerequests/index.blade.php
@@ -42,7 +42,7 @@
 					<td>{{ $changerequest->updated_at }}</td>
 					<td>{{ $changerequest->status }}</td>
 					<td>{{ $changerequest->comment }}</td>
-					{!! Form::open(array('class' => 'form-inline', 'method' => 'DELETE', 'route' => array('changerequests.destroy', $changerequest->id))) !!}
+					{!! Form::open(array('class' => 'form-inline', 'method' => 'DELETE', 'route' => array('changerequests.destroy', $changerequest->id), 'onsubmit' => 'return confirm(\'Are you sure to delete this source?\')')) !!}
 					<td>
 						{!! link_to_route('changerequests.edit', 'Review', array($changerequest->id), array('class' => 'btn btn-info btn-xs')) !!}
 						{!! Form::submit('Delete', array('class' => 'btn btn-danger btn-xs', 'style' => 'margin-left:3px;')) !!}
